修正楼梯自动跑采样秒数为double常量
SampleSeconds写成了float却用45d，Unity报CS0664编不过。改成double，和Editor计时一致。

+++ 3cDemo/Client/3C_Client/Assets/GameScripts/Main/Editor/CharacterPipeline/Diagnostics/GameplayLabFootIkKeyboardRouteDriver.cs
@@ -14,7 +14,7 @@ namespace ThirdPersonCharacter.Pipeline.Editor
     [InitializeOnLoad]
     public static class GameplayLabFootIkKeyboardRouteDriver
     {
-        const float SampleSeconds = 45d;
+        const double SampleSeconds = 45d;
 
         static bool s_Active;
         static GameplayLabFootIkStairAdPlan s_Plan;
